+++ test/src/Admin/Model/DoctrineRestServiceResourceFactoryTest.php
@@ -68,7 +68,8 @@ class DoctrineRestServiceResourceFactoryTest extends \PHPUnit_Framework_TestCase
             $this->container->has($dependency)->willReturn($presence);
         }
 
-        $this->setExpectedException(ServiceNotCreatedException::class, 'missing one or more dependencies');
+        $this->expectException(ServiceNotCreatedException::class);
+        $this->expectExceptionMessage('missing one or more dependencies');
         $factory($this->container->reveal());
     }